requirements.txt
openai
PyPDF2
python-docx
pytesseract
pillow
gspread
oauth2client
google-search-results
python-dotenv
